Ignoring failures when checking for hyperlinks
Adding LanguageServers.addAllSuccessful as an alternative to
LanguageServers.addAll so that exceptions from one of the requests being
combined can be ignored. Then using this new method in
OpenDeclarationHyperlinkDetector when combining the results from
definitions, declarations, typeDefinitions, and implementions because
some LSPs return errors instead of empty lists for some of these calls.
Instead, we combine only the successful results when building the list
of hyperlinks, ignoring the errors.

This should resolve https://github.com/eclipse-lsp4e/lsp4e/issues/1169